requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aldryn-apphooks-config==0.2.7  # via djangocms-blog
aldryn-boilerplates==0.7.4  # via aldryn-common, aldryn-video
aldryn-common==1.0.4      # via aldryn-search
aldryn-search==0.3.0      # via djangocms-blog
aldryn-video==1.0.0       # via nephila-widgets
beautifulsoup4==4.5.1     # via aldryn-video
click==6.6                # via pip-tools
cmsplugin-filer==1.1.3
contextlib2==0.5.4        # via raven
cssselect==0.9.2          # via premailer
cssutils==1.0.1           # via premailer
diff-match-patch==20121119  # via django-import-export
django-appconf==1.0.2     # via aldryn-boilerplates, aldryn-search, cmsplugin-filer, django-compressor, nephila-widgets
django-appdata==0.1.5     # via aldryn-apphooks-config, nephila-widgets
django-classy-tags==0.8.0
django-clear-cache==0.3
django-cms==3.3.3
django-compressor==2.1
django-environ==0.4.0
django-filer==1.2.5
django-formtools==1.0.0.1  # via django-cms
django-haystack==2.5.0    # via aldryn-search
django-import-export==0.5.1
django-meta-mixin==0.3.0  # via djangocms-blog
django-meta==1.3.1        # via django-meta-mixin, djangocms-blog, djangocms-page-meta
django-mptt==0.8.6        # via django-filer
django-parler==1.6.5      # via djangocms-blog
django-polymorphic==0.8.1  # via django-filer
django-redis==4.5.0
django-reversion==1.10.2
django-robots==2.0
django-sekizai==0.10.0
django-select2==4.3.2     # via djangocms-link
django-sortedm2m==1.3.2   # via aldryn-common
django-spurl==0.6.4       # via aldryn-search
django-standard-form==1.1.5  # via aldryn-search
django-taggit-autosuggest==0.3.0  # via djangocms-blog
django-taggit-templatetags==0.2.5  # via djangocms-blog
django-taggit==0.21.3     # via django-taggit-autosuggest, django-taggit-templatetags, djangocms-blog
django-templatetag-sugar==1.0  # via django-taggit-templatetags
django-treebeard==4.0.1
django==1.8.15
djangocms-admin-style==1.2.5
djangocms-apphook-setup==0.1.2  # via djangocms-blog
djangocms-attributes-field==0.1.1  # via cmsplugin-filer, djangocms-link, djangocms-video, nephila-widgets
djangocms-blog==0.8.8
djangocms-googlemap==0.5.2
djangocms-link==1.8.3
djangocms-page-meta==0.6.3
djangocms-text-ckeditor==3.2.1
djangocms-video==2.0.2
djmail==0.13
easy-thumbnails==2.3
ecdsa==0.13               # via paramiko
fabric3==1.12.post1
first==2.0.1              # via pip-tools
hiredis==0.2.0
html5lib==0.9999999
jsonfield==1.0.3          # via aldryn-video
lxml==3.6.4               # via aldryn-search, premailer
micawber==0.3.3           # via aldryn-video
paramiko==1.17.2          # via fabric3
pillow==3.4.1
pip-tools==1.7.0
premailer==3.0.1
psycopg2==2.6.2
pycrypto==2.6.1           # via paramiko
pytz==2016.7
raven==5.27.1
rcssmin==1.0.6            # via django-compressor
redis==2.10.5             # via django-redis
requests==2.11.1          # via premailer
rjsmin==1.0.12            # via django-compressor
six==1.10.0
tablib==0.11.2            # via django-import-export
unidecode==0.4.19         # via django-filer
urlobject==2.4.2          # via django-spurl
yurl==0.13                # via aldryn-boilerplates

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools                # via aldryn-search, django-polymorphic
